fix bug where first promise is not caught properly on connection
If `.catch` is called on a promise, a promise is generated and called. In this case if there was an error upon connection, you would get 2 errors. The first one would be the caught newly generated promise from the first promise (which is intended behavior, but on the wrong promise). The second is the uncaught yield of the assignment resulting in an UnhandledPromiseRejectionWarning.